% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_approximation.R
\name{alfers_dinges_approx}
\alias{alfers_dinges_approx}
\title{Normal approximation for Beta distribution}
\usage{
alfers_dinges_approx(Y)
}
\arguments{
\item{Y}{vector}
}
\value{
vector
}
\description{
Takes in a vector Y containing data from a beta distribution and returns data that is approximately normal.
Uses approximation found in "A Normal Approximation for Beta and Gamma Tail Probabilities"
(Alfers and Dinges)
}
\examples{
alfers_dinges_approx(rbeta(1000, 1, 10))
}
