% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{bind_data_frames}
\alias{bind_data_frames}
\title{Bind multiple (wide) data frames together}
\usage{
bind_data_frames(..., remove_na = TRUE)
}
\arguments{
\item{...}{list of data frames}

\item{remove_na}{remove columns with NA values? defualt = TRUE}
}
\value{
binded dataset
}
\description{
Bind multiple (wide) data frames together
}
\examples{
df1 <- data.frame(icgc_donor_id = 'DO1111',
                  A1BG = 0.4,
                  AAAS = 0.5)
df2 <- data.frame(icgc_donor_id = 'DO2222',
                  A1BG = 0.4,
                  AAAS = 0.5)
bind_data_frames(df1,df2)
}
