% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_converter.R
\name{convert_probe_to_gene}
\alias{convert_probe_to_gene}
\title{Convert probe ids to nearest genes}
\usage{
convert_probe_to_gene(
  probe_ids,
  icgc_donor_id,
  hm450 = NULL,
  chunk_size = 10000
)
}
\arguments{
\item{probe_ids}{vector}

\item{icgc_donor_id}{vector}

\item{hm450}{Large GRanges object}

\item{chunk_size}{size of chunks when converting probe ids to genes}
}
\value{
vector of mapped genes
}
\description{
Uses FDb.InfiniumMethylation.hg19 Bioconductor package to map probe ids to nearest genes in chunks
}
\examples{
library(FDb.InfiniumMethylation.hg19)
hm450 <- FDb.InfiniumMethylation.hg19::get450k()
convert_probe_to_gene(
  c("cg10861017", "cg10861599", "cg10861751", "cg10862535", "cg10862587"),
  "DO110234",
  hm450
)
}
