% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug}
\alias{debug}
\title{Debugger}
\usage{
debug(df, message)
}
\arguments{
\item{df}{current dataframe in the dplyr pipeline}

\item{message}{custom message used for debugging purposes}
}
\value{
the same dataframe to pipe into next dplyr function
}
\description{
Used to debug dplyr code by printing out custom messages
}
\examples{
df <- data.frame(x = seq(1:10),
                 y = rnorm(10))
debug(df, "debugging!")
}
