% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Mapping from project code to tissue}
\format{
A data frame with columns:
\describe{
\item{project_code}{name of the project}
\item{tissue}{the corresponding tissue }
}
}
\source{
ICGC portal
}
\usage{
dict
}
\description{
Mapping from project code to tissue
}
\keyword{datasets}
