% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{do_pca}
\alias{do_pca}
\title{Perform PCA}
\usage{
do_pca(data, transform = TRUE)
}
\arguments{
\item{data}{a dataframe with subjects/observations on the rows and variables on the columns}

\item{transform}{boolean indicating whether the first column needs to be converted to rownames, default is TRUE}
}
\value{
pca object
}
\description{
Perform PCA to reduce dimension of the dataset
}
\examples{
df <- data.frame(icgc_donor_id = c('DO10324', 'DO10486', 'DO10558', 'DO10631'),
                 A1BG = rnorm(4),
                 A2ML1 = rnorm(4),
                 A4GALT = rnorm(4),
                 A4GNT = rnorm(4))
do_pca(df)
}
