% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_probes.R
\name{filter_probes}
\alias{filter_probes}
\title{Filter probes based on mean absolute deviation}
\usage{
filter_probes(
  input,
  sep = ",",
  average_duplicate = TRUE,
  array_platform = "HumanMethylation450_after_2011_08_02",
  num_probes = 5000,
  cols_to_read = readr::cols_only(icgc_donor_id = "c", probe_id = "c",
    methylation_value = "d", array_platform = "c")
)
}
\arguments{
\item{input}{data frame with icgc_donor_id, probe_id and methylation_value columns, or path to file}

\item{sep}{delimiter for data_file, default = ','}

\item{average_duplicate}{average methylation value for duplicate (probe_id, methylation_value) pairs, default=FALSE}

\item{array_platform}{array platform to filter on}

\item{num_probes}{number of probes to subset}

\item{cols_to_read}{cols_only object indicating which columns to read if input is a file path,
default = cols_only(icgc_donor_id ='c', probe_id = 'c', methylation_value = 'd')}
}
\value{
long data frame subsetted on probes
}
\description{
Processes dataset and returns a subset based on most variable probes
}
\examples{
input <- data.frame(icgc_donor_id = c('DO1000','DO2000'),
                    probe_id = c('cg10861017','cg10861017'),
                    methylation_value = c(0.74, 0.92),
                    array_platform = rep('HumanMethylation450_after_2011_08_02', 2))
filter_probes(input)
}
