% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_converter.R
\name{get_genes}
\alias{get_genes}
\title{Helper function to convert_probe_to_gene}
\usage{
get_genes(probe_ids, hm450)
}
\arguments{
\item{probe_ids}{list of probe ids}

\item{hm450}{Large GRanges object}
}
\value{
vector of mapped genes
}
\description{
Helper function to convert_probe_to_gene
}
\examples{
library(FDb.InfiniumMethylation.hg19)
hm450 <- FDb.InfiniumMethylation.hg19::get450k()
get_genes(c("cg10861017", "cg10861599", "cg10861751"), hm450)
}
