% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_approximation.R
\name{get_normalization_method}
\alias{get_normalization_method}
\title{Returns corresponding normalization method}
\usage{
get_normalization_method(name)
}
\arguments{
\item{name}{string}
}
\value{
function
}
\description{
Given a string, returns the corresponding normalization function
}
\examples{
get_normalization_method("alfers_dinges")
}
