% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_tissue_locations}
\alias{get_tissue_locations}
\title{Map donor ids to tissue locations using project code}
\usage{
get_tissue_locations(input, sep = NA, icgc_donor_ids)
}
\arguments{
\item{input}{donor dataset that maps icgc donor ids to project codes or file name
(only requires icgc_donor_id and project_code columns)}

\item{sep}{character, delimiter if input is file name}

\item{icgc_donor_ids}{dataframe with the relevant donor ids}
}
\value{
dataframe with the mapped tissues
}
\description{
Map donor ids to tissue locations using project code
}
\examples{
input <- data.frame(icgc_donor_id = c('DO1111', 'DO2222'),
                    project_code = c('BRCA-US', 'OV-AU'))
icgc_donor_ids <- data.frame(icgc_donor_id = c('DO1111', 'DO2222'))
get_tissue_locations(input=input, icgc_donor_ids=icgc_donor_ids)
}
