% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load_data}
\alias{load_data}
\title{User friendly data loader}
\usage{
load_data(input, sep, cols_to_read = NULL)
}
\arguments{
\item{input}{can be a file name or a data frame}

\item{sep}{delimiter if input is a file name}

\item{cols_to_read}{cols_only object indicating which columns to read if input is a file path, default = NULL}
}
\value{
loaded data
}
\description{
User friendly data loader
}
\examples{
load_data(data.frame(x = c(1,2,3),
                     y = c(4,5,6)))
}
