% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{methylation_pipeline}
\alias{methylation_pipeline}
\title{Entire preprocessing pipeline. First filters probes then does probe to gene conversion}
\usage{
methylation_pipeline(
  input,
  filename = NA,
  sep = "\\t",
  average_duplicate = TRUE,
  array_platform = "HumanMethylation450_after_2011_08_02",
  num_probes = 5000,
  normalization_method = "alfers_dinges",
  chunk_size = 10000,
  cols_to_read_probe = readr::cols_only(icgc_donor_id = "c", probe_id = "c",
    methylation_value = "d", array_platform = "c"),
  cols_to_read_meth = readr::cols_only(icgc_donor_id = "c", probe_id = "c",
    methylation_value = "d"),
  tissue_input = "donor.tsv",
  tissue_sep = "\\t"
)
}
\arguments{
\item{input}{data frame with icgc_donor_id, probe_id and methylation_value columns, or path to file}

\item{filename}{filename for writing matrix to .csv file}

\item{sep}{delimiter for input, default = '\t'}

\item{average_duplicate}{average methylation value for duplicate (probe_id, methylation_value) pairs, default=FALSE}

\item{array_platform}{array platform to filter on}

\item{num_probes}{number of probes to subset}

\item{normalization_method}{method to normalize methylation values, default = 'alfers_dinges'}

\item{chunk_size}{size of chunks when converting probe ids to genes}

\item{cols_to_read_probe}{cols_only object indicating which columns to read if input is a file path,
default = cols_only(icgc_donor_id ='c', probe_id = 'c', methylation_value = 'd')}

\item{cols_to_read_meth}{cols_only object indicating which columns to read if input is a file path,
default = cols_only(icgc_donor_id ='c', probe_id = 'c', methylation_value = 'd')}

\item{tissue_input}{data frame with icgc_donor_id, probe_id and methylation_value columns, or path to file}

\item{tissue_sep}{delimiter if tissue_input is a filename}
}
\value{

}
\description{
Entire preprocessing pipeline. First filters probes then does probe to gene conversion
}
\examples{
input <- data.frame(icgc_donor_id = c('DO1000','DO2000'),
                    probe_id = c('cg10861017','cg10861017'),
                    methylation_value = c(0.74, 0.92),
                    array_platform = rep('HumanMethylation450_after_2011_08_02', 2))

tissue_input <- data.frame(icgc_donor_id = c('DO1111', 'DO2222'),
                           project_code = c('BRCA-US', 'OV-AU'))

methylation_pipeline(input=input,
                     tissue_input=tissue_input)
}
