% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_na.R
\name{no_na}
\alias{no_na}
\title{Checks for NA values in a vector}
\usage{
no_na(vec)
}
\arguments{
\item{vec}{vector}
}
\value{
boolean
}
\description{
If there are any NA values in the vector, the function will return FALSE. Otherwise, it will return TRUE.
If there are any errors, the function will return FALSE
}
\examples{
no_na(c(1,2,3,NA,4,5)) # Returns FALSE
no_na(c(1,2,3,4,5,6)) # Returns TRUE
}
