% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_approximation.R
\name{peizer_pratt_approx}
\alias{peizer_pratt_approx}
\title{Normal approximation for Beta distribution}
\usage{
peizer_pratt_approx(Y)
}
\arguments{
\item{Y}{vector}
}
\value{
vector
}
\description{
Takes in a vector Y containing data from a beta distribution and returns data that is approximately normal.
Uses approximation found in "A Normal Approximation for Binomial, F, Beta, and Other Common, Related Tail Probabilities, I"
(Peizer and Pratt)
}
\examples{
peizer_pratt_approx(rbeta(1000, 1, 10))
}
