% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{process_raw_methylation_data}
\alias{process_raw_methylation_data}
\title{Preprocesses raw methylation data}
\usage{
process_raw_methylation_data(
  input,
  sep = ",",
  normalization_method = "alfers_dinges",
  chunk_size = 10000,
  average_duplicate = FALSE,
  cols_to_read = cols_only(icgc_donor_id = "c", probe_id = "c", methylation_value =
    "d")
)
}
\arguments{
\item{input}{data frame with icgc_donor_id, probe_id and methylation_value columns, or path to file}

\item{sep}{delimiter for data_file, default = ','}

\item{normalization_method}{method to normalize methylation values, default = 'alfers_dinges'}

\item{chunk_size}{size of chunks when converting probe ids to genes}

\item{average_duplicate}{average methylation value for duplicate (probe_id, methylation_value) pairs, default=FALSE}

\item{cols_to_read}{cols_only object indicating which columns to read if input is a file path,
default = cols_only(icgc_donor_id ='c', probe_id = 'c', methylation_value = 'd')}
}
\value{
list object: data.frame matrix whose rows are donors and columns are genes, gene dataframe, donor id dataframe
}
\description{
Process methylation data (in long form) from icgc portal through pipeline and outputs a list (methylation_matrix, donor_ids, gene_ids)
methylation_matrix - donors on rows, genes on columns, methyaltion values as elements of matrix
donor_ids - vector of donor ids for rows of methylation matrix
gene_ids - vector of genes for columns of methylation matrix
}
\examples{
data <- data.frame(icgc_donor_id = c('DO1000'),
                   probe_id = 'cg10861017',
                   methylation_value = 0.74)
process_raw_methylation_data(data, 'alfers_dinges')
}
