% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_approximation.R
\name{wise_approx}
\alias{wise_approx}
\title{Normal approximation for Beta distribution}
\usage{
wise_approx(Y)
}
\arguments{
\item{Y}{vector}
}
\value{
vector
}
\description{
Takes in a vector Y containing data from a beta distribution and returns data that is approximately normal.
Uses Wise's transformation: (-logY)^(1/3)
}
\examples{
wise_approx(rbeta(1000, 1, 10))
}
